requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain
langchain-core
unstructured 
openpyxl
sentence-transformers 
faiss-cpu
InstructorEmbedding
langgraph
grandalf
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
typing-inspect==0.9.0
typing_extensions==4.12.2
langchain-tavily